+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"integration_test","path":"/home/jay-flow/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/integration_test-0.8.1/","dependencies":[]}],"android":[{"name":"integration_test","path":"/home/jay-flow/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/integration_test-0.8.1/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"integration_test","dependencies":[]}],"date_created":"2022-01-10 18:16:09.760384","version":"2.8.1"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"integration_test","path":"/home/jay-flow/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/integration_test-0.8.1/","dependencies":[]}],"android":[{"name":"integration_test","path":"/home/jay-flow/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/integration_test-0.8.1/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"integration_test","dependencies":[]}],"date_created":"2022-01-10 20:38:52.561126","version":"2.8.1"}
