+++ android/app/src/main/res/drawable-v21/launch_background.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!-- Modify this file to customize your launch splash screen -->
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="?android:colorBackground" />
+    <item android:drawable="@android:color/white" />
 
     <!-- You can insert your own image assets here -->
     <item>
-         <bitmap
+        <bitmap
             android:gravity="center"
             android:src="@mipmap/ic_launcher" />
     </item>
