+++ lib/l10n/app_en.arb
@@ -1,3 +1,4 @@
 {
-    "doobooLab": "dooboolab"
+    "appName": "dooboolab",
+    "helloWorld": "Hello world"
 }
