+++ lib/l10n/app_ko.arb
@@ -1,3 +1,4 @@
 {
-    "doobooLab": "두부랩"
+    "appName": "두부랩",
+    "helloWorld": "헬로월드"
 }
