+++ linux/CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.10)
 project(runner LANGUAGES CXX)
 
 set(BINARY_NAME "flutter_boilerplate")
-set(APPLICATION_ID "com.example.flutter_boilerplate")
+set(APPLICATION_ID "com.dooboolab.flutter_boilerplate")
 
 cmake_policy(SET CMP0063 NEW)
